Move ChunkFinder to new file
